[WIP] Major refactor. Renamed jobs to automate
