+++ aero/models/provenance.py
@@ -9,8 +9,8 @@ from sqlalchemy import String
 
 from aero.app import db
 
-from aero.jobs.user_flow import run_flow
-from aero.jobs.timer import set_timer
+from aero.automate.user_flow import run_flow
+from aero.automate.timer import set_timer
 from aero.globus.utils import FlowEnum
 
 provenance_derivation = db.Table(
